requirements.txt
python-telegram-bot==22.0
requests==2.32.3
email_validator==2.2.0
python-dotenv==1.1.0
requests==2.32.3

